# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EGH Panel API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /nodes:
    get:
      operationId: listNodes
      tags: [nodes]
      summary: List all nodes (admin only)
      responses:
        "200":
          description: List of nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Node"
    post:
      operationId: createNode
      tags: [nodes]
      summary: Create a node (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNodeBody"
      responses:
        "201":
          description: Node created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
  /nodes/{id}:
    get:
      operationId: getNode
      tags: [nodes]
      summary: Get node by ID (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Node details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateNode
      tags: [nodes]
      summary: Update node (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNodeBody"
      responses:
        "200":
          description: Node updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
    delete:
      operationId: deleteNode
      tags: [nodes]
      summary: Delete node (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Allocations
  /nodes/{nodeId}/allocations:
    get:
      operationId: listAllocations
      tags: [allocations]
      summary: List allocations for a node (admin only)
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of allocations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Allocation"
    post:
      operationId: createAllocation
      tags: [allocations]
      summary: Create allocation for a node (admin only)
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAllocationBody"
      responses:
        "201":
          description: Allocation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allocation"
  /allocations/{id}:
    delete:
      operationId: deleteAllocation
      tags: [allocations]
      summary: Delete allocation (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Servers (Admin)
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Node:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        fqdn:
          type: string
        scheme:
          type: string
          enum: [http, https]
        daemonPort:
          type: integer
        locationId:
          type: ["integer", "null"]
        locationName:
          type: ["string", "null"]
        isPublic:
          type: boolean
        memoryTotal:
          type: integer
        memoryOverallocate:
          type: integer
        diskTotal:
          type: integer
        diskOverallocate:
          type: integer
        status:
          type: string
          enum: [online, offline, maintenance]
        serverCount:
          type: integer
        allocationCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - fqdn
        - scheme
        - daemonPort
        - isPublic
        - memoryTotal
        - memoryOverallocate
        - diskTotal
        - diskOverallocate
        - status
        - serverCount
        - allocationCount
        - createdAt
    NodeDetail:
      allOf:
        - $ref: "#/components/schemas/Node"
        - type: object
          properties:
            allocations:
              type: array
              items:
                $ref: "#/components/schemas/Allocation"
    CreateNodeBody:
      type: object
      properties:
        name:
          type: string
        fqdn:
          type: string
        scheme:
          type: string
          enum: [http, https]
        daemonPort:
          type: integer
        isPublic:
          type: boolean
        memoryTotal:
          type: integer
        memoryOverallocate:
          type: integer
        diskTotal:
          type: integer
        diskOverallocate:
          type: integer
      required:
        - name
        - fqdn
        - scheme
        - daemonPort
        - memoryTotal
        - diskTotal
    Allocation:
      type: object
      properties:
        id:
          type: integer
        nodeId:
          type: integer
        ip:
          type: string
        port:
          type: integer
        alias:
          type: ["string", "null"]
        isAssigned:
          type: boolean
        serverId:
          type: ["integer", "null"]
        serverName:
          type: ["string", "null"]
      required:
        - id
        - nodeId
        - ip
        - port
        - isAssigned
    CreateAllocationBody:
      type: object
      properties:
        ip:
          type: string
        ports:
          type: array
          items:
            type: integer
        alias:
          type: string
      required:
        - ip
        - ports
...